% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function.R
\name{check_directlfq_Module}
\alias{check_directlfq_Module}
\title{Check the directlfq environment}
\usage{
check_directlfq_Module(confirm_or_not = T)
}
\arguments{
\item{confirm_or_not}{Whether provide a second comfirm.}
}
\value{
TURE if the environment is fine, and FALSE if the environment is missed.
}
\description{
This function check and install (if absent) a python conda environment contains the directlfq library
}
\examples{
Rdirectlfq::check_directlfq_Module()
}
