% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directlfq.R
\name{directlfq}
\alias{directlfq}
\title{Run directlfq}
\usage{
directlfq(df, ncores = 4, temp_dir = tempdir(), reformatted = T, ...)
}
\arguments{
\item{df}{A dataframe in the generic input format of directlfq with the following columns: "protein", "ion", runs.}

\item{ncores}{Use howmany cores, default is 4. If it is 0, will use all available threads.}

\item{reformatted}{If TRUE means df is in generic input format, else df is a result table from quantity.
We recommend use a formatted input}

\item{...}{To be completed in the future. Other parameters passed to directlfq.}

\item{tempdir}{The path of temp files.}
}
\value{
A dataframe contains the directlfq result, protein intensities.
}
\description{
This function perform directlfq via the [directlfq]("https://github.com/MannLabs/directlfq/") module.
}
\details{
Specifically, this R package installs a conda environment containing the directlfq library using reticulate.
Then, directlfq calculations are performed by calling Python scripts within that environment.
This function will try to
Or, you can set up the required conda environment by running the check_directlfq_Module() function.
If you are a Windows user, you may need to run this function with administrator privileges.
The directlfq function will also attempt to deploy this conda environment if it is missing.
}
\examples{
## A diann report.tsv
data(example_diann_res)
pro_int = directlfq(example_diann_res,reformatted = F)

## A generic input table, which has been reformatted.
data(example_generic_input)
pro_int = directlfq(example_generic_input, reformatted = T)

}
